remove public from Analytics class
