+++ Client/KarhooAnalyticsServiceWithNotifications.swift
